Rename `persistence/remote` to `remotestream`.
