+++ persistence/remotestream/doc.go
@@ -1,0 +1,3 @@
+// Package remotestream contains a persistence.Stream implementation that
+// streams messages via the messaging API's EventStream service.
+package remotestream
